# specs/001-cleaning-company-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Nuqaa Asir Cleaning Booking & Operations Platform API
  version: "1.0.0"
  description: >
    REST API for the customer booking portal and admin operations dashboard.
    Two roles only: CUSTOMER and ADMIN. All endpoints are under /api/v1.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /addresses/me:
    get:
      tags: [Addresses]
      summary: List own addresses
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Address" } } } } } }
    post:
      tags: [Addresses]
      summary: Add an address
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/AddressInput" } } } }
      responses: { "201": { description: Created } }
  /addresses/me/{id}:
    patch:
      tags: [Addresses]
      summary: Update own address
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/AddressInput" } } } }
      responses: { "200": { description: Updated } }
    delete:
      tags: [Addresses]
      summary: Remove own address
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "204": { description: Deleted } }
  /customers/{customerId}/addresses:
    get:
      tags: [Addresses]
      summary: List a customer's addresses (Admin)
      parameters:
        - in: path
          name: customerId
          required: true
          schema: { type: string, format: uuid }
      responses: { "200": { description: OK } }

  # ---------------------------------------------------------------- Catalog
...
  /bookings:
    get:
      tags: [Bookings]
      summary: List bookings (Customer sees own; Admin sees all, filterable)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/BookingStatus" }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/BookingListResponse" } } }
    post:
      tags: [Bookings]
      summary: Create a booking (Customer, self-service web flow)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingCreateRequest" }
        description: Idempotency-Key header required to prevent duplicate submissions
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Booking created
          content: { application/json: { schema: { $ref: "#/components/schemas/Booking" } } }
        "409": { description: Service area/time slot no longer available }
        "422": { $ref: "#/components/responses/ValidationError" }
  /bookings/admin:
    post:
      tags: [Bookings]
      summary: Create a booking on behalf of a phone/WhatsApp caller (Admin)
      requestBody:
        content: { application/json: { schema: { $ref: "#/components/schemas/AdminBookingCreateRequest" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Booking" } } } }
  /bookings/reference/{referenceNumber}:
    get:
      tags: [Bookings]
      summary: Public booking lookup by reference + verification token
      security: []
      parameters:
        - in: path
          name: referenceNumber
          required: true
          schema: { type: string }
        - in: query
          name: token
          required: true
          schema: { type: string }
          description: unguessable verification token issued at booking creation (FR-077)
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/BookingSummary" } } } }
        "403": { description: Token missing or invalid }
        "404": { $ref: "#/components/responses/NotFound" }
  /bookings/{id}:
    get:
      tags: [Bookings]
      summary: Get booking detail (owner or Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Booking" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Bookings]
      summary: Edit booking details prior to confirmation (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: Updated } }
  /bookings/{id}/confirm:
    post:
      tags: [Bookings]
      summary: Confirm a pending booking, optionally overriding price (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                priceOverride: { type: integer, nullable: true, description: minor units }
                priceOverrideReason: { type: string, nullable: true }
      responses:
        "200": { description: Confirmed }
        "409": { description: Missing required fields (FR-034) }
  /bookings/{id}/reject:
    post:
      tags: [Bookings]
      summary: Reject a pending booking (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      requestBody:
        content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string } } } } }
      responses: { "200": { description: Rejected } }
  /bookings/{id}/schedule:
    post:
      tags: [Bookings]
      summary: Set planned start/end time and optional internal handling note (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [scheduledStartAt, scheduledEndAt]
              properties:
                scheduledStartAt: { type: string, format: date-time }
                scheduledEndAt: { type: string, format: date-time }
                internalHandlingNote: { type: string, nullable: true }
                overrideCapacity: { type: boolean, default: false }
      responses:
        "200": { description: Scheduled }
        "409": { description: Time-slot capacity exceeded and overrideCapacity=false (FR-030) }
  /bookings/{id}/reschedule:
    post:
      tags: [Bookings]
      summary: Reschedule to a new date/time (Admin, or Customer request routed to Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [newDate, reason]
              properties:
                newDate: { type: string, format: date }
                newTimeSlotId: { type: string, format: uuid, nullable: true }
                reason: { type: string }
      responses: { "200": { description: Rescheduled } }
  /bookings/{id}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel a booking (Customer own booking, or Admin any booking) - always free, no fee (FR-040)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      requestBody:
        content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string } } } } }
      responses: { "200": { description: Cancelled } }
  /bookings/{id}/en-route:
    post:
      tags: [Bookings]
      summary: Mark en route (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: Marked en route } }
  /bookings/{id}/arrive:
    post:
      tags: [Bookings]
      summary: Mark arrived (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: Marked arrived } }
  /bookings/{id}/start:
    post:
      tags: [Bookings]
      summary: Mark job started; transitions status to IN_PROGRESS (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses:
        "200": { description: Started }
        "409": { description: Arrival not yet recorded (FR-036) }
  /bookings/{id}/complete:
    post:
      tags: [Bookings]
      summary: Mark job complete; requires checklist review passed (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses:
        "200": { description: Completed; queues customer feedback notification }
        "409": { description: Required checklist items unfinished (FR-048, FR-051) }
  /bookings/{id}/history:
    get:
      tags: [Bookings]
      summary: Booking status history (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/BookingStatusHistoryEntry" } } } }

  # ---------------------------------------------------------------- Checklists
...
  /commercial-accounts/{id}:
    patch:
      tags: [Commercial]
      summary: Update commercial account (Admin)
      parameters: [ { $ref: "#/components/parameters/IdParam" } ]
      responses: { "200": { description: Updated } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Forbidden:
      description: Authenticated but not permitted (e.g., Customer accessing another customer's booking, or Customer on an Admin-only route)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: BOOKING_SLOT_UNAVAILABLE }
            message: { type: string }
            fieldErrors: { type: object, additionalProperties: { type: array, items: { type: string } } }
            requestId: { type: string }
    BookingStatus:
      type: string
      enum: [DRAFT, PENDING, CONFIRMED, RESCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED, REJECTED, COMPLAINT_OPENED]
...
    Address:
      type: object
      properties:
        id: { type: string, format: uuid }
        label: { type: string, nullable: true }
        city: { type: string }
        neighborhood: { type: string }
        street: { type: string, nullable: true }
        buildingNumber: { type: string, nullable: true }
        unitNumber: { type: string, nullable: true }
        landmark: { type: string, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        mapUrl: { type: string, nullable: true }
        serviceAreaId: { type: string, format: uuid }
        isDefault: { type: boolean }
    AddressInput:
      allOf:
        - $ref: "#/components/schemas/Address"
      required: [city, neighborhood, serviceAreaId]
...
    BookingCreateRequest:
      type: object
      required: [quoteId, addressId, propertyType, preferredDate, contactName, contactPhone, consentAccepted]
      properties:
        quoteId: { type: string, format: uuid }
        addressId: { type: string, format: uuid }
        propertyType: { type: string, enum: [APARTMENT, VILLA, OFFICE, SHOP, CLINIC, FURNISHED_UNIT, OTHER] }
        propertyDetails: { type: object }
        preferredDate: { type: string, format: date }
        preferredTimeSlotId: { type: string, format: uuid, nullable: true }
        contactName: { type: string }
        contactPhone: { type: string }
        customerNotes: { type: string, nullable: true }
        consentAccepted: { type: boolean }
    AdminBookingCreateRequest:
      allOf:
        - $ref: "#/components/schemas/BookingCreateRequest"
      properties:
        customerId: { type: string, format: uuid, nullable: true, description: existing customer, or omit to create a new INVITED customer }
        newCustomer:
          type: object
          nullable: true
          properties:
            fullName: { type: string }
            phone: { type: string }
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        referenceNumber: { type: string }
        status: { $ref: "#/components/schemas/BookingStatus" }
        customerId: { type: string, format: uuid }
        addressId: { type: string, format: uuid }
        propertyType: { type: string }
        scheduledStartAt: { type: string, format: date-time, nullable: true }
        scheduledEndAt: { type: string, format: date-time, nullable: true }
        internalHandlingNote: { type: string, nullable: true }
        enRouteAt: { type: string, format: date-time, nullable: true }
        arrivedAt: { type: string, format: date-time, nullable: true }
        startedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        totalSnapshot: { type: integer }
        currency: { type: string }
        createdAt: { type: string, format: date-time }
    BookingSummary:
      type: object
      description: Reduced-PII view returned by the public reference lookup endpoint
      properties:
        referenceNumber: { type: string }
        status: { $ref: "#/components/schemas/BookingStatus" }
        scheduledStartAt: { type: string, format: date-time, nullable: true }
        serviceName: { type: string }
    BookingListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Booking" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    BookingStatusHistoryEntry:
      type: object
      properties:
        fromStatus: { $ref: "#/components/schemas/BookingStatus" }
        toStatus: { $ref: "#/components/schemas/BookingStatus" }
        reason: { type: string, nullable: true }
        actorUserId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
...